# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Diet Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tracker/log-meal:
    post:
      operationId: logMeal
      tags: [tracker]
      summary: Log a meal entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogMealBody"
      responses:
        "201":
          description: Meal logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealEntry"
  /tracker/log-water:
    post:
      operationId: logWater
      tags: [tracker]
      summary: Log water intake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogWaterBody"
      responses:
        "200":
          description: Water logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterEntry"
  /tracker/meals:
    get:
      operationId: listMealLogs
      tags: [tracker]
      summary: List meal logs for a date range
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Meal logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MealEntry"
  /tracker/weight:
    post:
      operationId: logWeight
      tags: [tracker]
      summary: Log a weight entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogWeightBody"
      responses:
        "201":
          description: Weight logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeightEntry"
    get:
      operationId: listWeightLogs
      tags: [tracker]
      summary: List weight logs
      responses:
        "200":
          description: Weight log history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightEntry"
...
components:
  schemas:
...
    LogMealBody:
      type: object
      properties:
        mealType:
          type: string
          enum: [breakfast, lunch, dinner, snack]
        foodName:
          type: string
        portionSize:
          type: string
        calories:
          type: integer
        proteinGrams:
          type: number
        carbsGrams:
          type: number
        fatGrams:
          type: number
        loggedAt:
          type: string
          format: date
      required:
        - mealType
        - foodName
        - portionSize
        - calories
    MealEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        mealType:
          type: string
        foodName:
          type: string
        portionSize:
          type: string
        calories:
          type: integer
        proteinGrams:
          type: number
        carbsGrams:
          type: number
        fatGrams:
          type: number
        loggedAt:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - mealType
        - foodName
        - portionSize
        - calories
        - loggedAt
        - createdAt
    LogWaterBody:
      type: object
      properties:
        amountMl:
          type: integer
        loggedAt:
          type: string
          format: date
      required:
        - amountMl
    WaterEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amountMl:
          type: integer
        loggedAt:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amountMl
        - loggedAt
        - createdAt
    LogWeightBody:
      type: object
      properties:
        weightKg:
          type: number
        loggedAt:
          type: string
          format: date
      required:
        - weightKg
    WeightEntry:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        weightKg:
          type: number
        loggedAt:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - weightKg
        - loggedAt
        - createdAt
...