# lib/api-zod/src/generated/types/logMealBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Diet Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { LogMealBodyMealType } from "./logMealBodyMealType";

export interface LogMealBody {
  mealType: LogMealBodyMealType;
  foodName: string;
  portionSize: string;
  calories: number;
  proteinGrams?: number;
  carbsGrams?: number;
  fatGrams?: number;
  loggedAt?: Date;
}