# lib/api-zod/src/generated/types/weightEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Diet Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface WeightEntry {
  id: number;
  userId: number;
  weightKg: number;
  loggedAt: Date;
  createdAt: Date;
}